+++ wave_reader.cpp
@@ -34,7 +34,6 @@
 #include "wave_reader.h"
 #include "error_reporter.h"
 #include <sstream>
-#include <string.h>
 
 using std::ostringstream;
 using std::exception;
@@ -56,7 +55,7 @@ WaveReader::WaveReader(string filename) :
     }
 
     ifs.seekg(0, ifs.end);
-    fileSize = ifs.tellg();
+    fileSize = (unsigned int) ifs.tellg();
     ifs.seekg(0, ifs.beg);
 
     try {
@@ -111,7 +110,7 @@ WaveReader::WaveReader(string filename) :
         throw error;
     }
 
-    dataOffset = ifs.tellg();
+    dataOffset = (unsigned int) ifs.tellg();
 }
 
 WaveReader::~WaveReader()
